started refactor for redux
