+++ server-list-app/.babelrc
@@ -1,4 +1,7 @@
 {
     "presets": ["@babel/env", "@babel/preset-react"],
-    "plugins": ["@babel/plugin-proposal-class-properties"]
+    "plugins": [
+        "@babel/plugin-proposal-class-properties", 
+        "transform-object-assign"
+    ]
 }
