+++ server-list-app/package.json
@@ -16,6 +16,8 @@
     "@babel/preset-env": "^7.1.0",
     "@babel/preset-react": "^7.0.0",
     "babel-loader": "^8.0.2",
+    "babel-plugin-transform-object-assign": "^6.22.0",
+    "babel-polyfill": "^6.26.0",
     "css-loader": "^1.0.0",
     "file-loader": "^4.2.0",
     "style-loader": "^0.23.0",
@@ -24,9 +26,12 @@
     "webpack-dev-server": "^3.1.8"
   },
   "dependencies": {
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
-    "react-hot-loader": "^4.3.11",
-    "react-router-dom": "^5.1.1"
+    "cross-fetch": "^3.0.4",
+    "react": "^16.10.1",
+    "react-dom": "^16.10.1",
+    "react-redux": "^7.1.1",
+    "react-router-dom": "^5.1.1",
+    "redux": "^4.0.4",
+    "redux-thunk": "^2.3.0"
   }
 }
