+++ server-list-app/src/Login/State/Actions.js
@@ -1,0 +1,17 @@
+
+
+export class LoginActionTypes {
+  static LOGIN_VALIDATION_PROGRESS_STATE = Symbol();
+  static LOGIN_VALIDATION_HAS_ERRORS = Symbol();
+}
+
+export class LoginActions {
+  static setValidationProgressState(progressCurrentState) {
+    return {
+      type: LoginActionTypes.LOGIN_VALIDATION_PROGRESS_STATE,
+      progressCurrentState: progressCurrentState
+    };
+  }
+}
+
+export default LoginActions;
