+++ server-list-app/src/Utils/AuthenticationAPI.js
@@ -1,4 +1,5 @@
 import APIUrls from "./APIUrls.js"
+import fetch from "cross-fetch"
 
 class AuthenticationAPI {
