+++ server-list-app/src/reducers.js
@@ -1,0 +1,6 @@
+import { combineReducers } from "redux";
+import LoginReducer from "./Login/State/Reducer.js"
+
+export default combineReducers({
+    LoginReducer
+  });
